fix(yg_radar): fix compilation issue of sign-compare

+++ modules/drivers/radar/yg_radar/yg_radar_message_manager.cc
@@ -41,8 +41,10 @@ void YgRadarMessageManager::Parse(const uint32_t message_id,
                                   const uint8_t* data, int32_t length) {
   // Parse method triggered by CanReceiver when a new message is received
 
-  if (message_id > Radarobjectlist560::ID + message_id_offset_ ||
-      message_id < Radarobjectinfo540::ID + message_id_offset_) {
+  if (message_id >
+          static_cast<uint32_t>(Radarobjectlist560::ID + message_id_offset_) ||
+      message_id <
+          static_cast<uint32_t>(Radarobjectinfo540::ID + message_id_offset_)) {
     // ignore messages that not match the message_id_offset_
     return;
   }
@@ -60,7 +62,7 @@ void YgRadarMessageManager::Parse(const uint32_t message_id,
     protocol_data->Parse(data, length, &sensor_data_);
     // object list received
     if (((message_id - message_id_offset_) & 0xFFE0) ==
-        Radarobjectlist560::ID) {
+        static_cast<uint32_t>(Radarobjectlist560::ID)) {
       auto& obj_list = sensor_data_.radar_object_list_560();
       current_obstacle_size_ = obj_list.object_number();
       // trigger obstacle message publish
@@ -79,7 +81,7 @@ void YgRadarMessageManager::Parse(const uint32_t message_id,
     }
     // object info received
     if (((message_id - message_id_offset_) & 0xFFE0) ==
-        Radarobjectinfo540::ID) {
+        static_cast<uint32_t>(Radarobjectinfo540::ID)) {
       auto& obj_info = sensor_data_.radar_object_info_540();
       auto& obstacle =
           (*(radar_obstacles_
